+++ Algorithm.CSharp/OptionStrategyFilteringUniversePutButterflyRegressionAlgorithm.cs
@@ -19,6 +19,7 @@ using System.Collections.Generic;
 using System.Linq;
 using QuantConnect.Data.Market;
 using QuantConnect.Securities.Option;
+using QuantConnect.Securities.Option.StrategyMatcher;
 
 namespace QuantConnect.Algorithm.CSharp
 {
@@ -69,8 +70,11 @@ namespace QuantConnect.Algorithm.CSharp
                 throw new Exception($"No contract returned match condition");
             }
 
-            var strategy = OptionStrategies.PutButterfly(_optionSymbol, itmStrike, atmStrike, otmStrike, expiry);
-            Buy(strategy, 1);
+            MarketOrder(otmContract.Symbol, +1);
+            MarketOrder(atmContract.Symbol, -2);
+            MarketOrder(itmContract.Symbol, +1);
+
+            AssertOptionStrategyIsPresent(OptionStrategyDefinitions.ButterflyPut.Name, 1);
         }
 
         /// <summary>
