+++ Algorithm.CSharp/OptionStrategyFilteringUniverseStraddleRegressionAlgorithm.cs
@@ -19,6 +19,7 @@ using System.Collections.Generic;
 using System.Linq;
 using QuantConnect.Data.Market;
 using QuantConnect.Securities.Option;
+using QuantConnect.Securities.Option.StrategyMatcher;
 
 namespace QuantConnect.Algorithm.CSharp
 {
@@ -61,8 +62,10 @@ namespace QuantConnect.Algorithm.CSharp
                 throw new Exception($"No contract returned match condition");
             }
 
-            var strategy = OptionStrategies.Straddle(_optionSymbol, strike, expiry);
-            Buy(strategy, 1);
+            MarketOrder(callContract.Symbol, +1);
+            MarketOrder(putContract.Symbol, +1);
+
+            AssertOptionStrategyIsPresent(OptionStrategyDefinitions.Straddle.Name, 1);
         }
 
         /// <summary>
